refactor: Use nullptr instead of NULL.
- Make Server::run() virtual and override it instead of hiding it.
- Clean up some minor things.

+++ shttps/Connection.h
@@ -313,7 +313,7 @@ namespace shttps {
         bool _chunked_transfer_out; //!< output data is sent in chunks
         bool _finished;             //!< Transfer of response data finished
         char *_content;             //!< Content if content-type is "text/plain", "application/json" etc.
-        unsigned content_length;    //!< length of body in octets (used if not chunked transfer)
+        size_t content_length;    //!< length of body in octets (used if not chunked transfer)
         std::string _content_type;   //!< Content-type (mime type of content)
         std::ofstream *cachefile;   //!< pointer to cache file
         char *outbuf;               //!< If not NULL, pointer to the output buffer (buffered output used)
